+++ neuroph-2.9/Core/src/main/java/org/neuroph/eval/classification/ConfusionMatrix.java
@@ -99,7 +99,7 @@ public class ConfusionMatrix {
     }
 
     public int getTruePositive(int clsIdx) {
-        return (int) values[clsIdx][clsIdx];
+        return values[clsIdx][clsIdx];
     }
 
     public int getTrueNegative(int clsIdx) {
